All floats are now float64

+++ control/control.go
@@ -5,7 +5,7 @@ import "dhemery.com/panelgen/shape"
 type Frame interface {
 	shape.Bounded
 	// At returns a copy of the frame translated by dx, dy
-	Translate(dx, dy float32) Frame
+	Translate(dx, dy float64) Frame
 	// Svg returns a shape.Svg with the same bounds and content as the frame
 	Svg() shape.Svg
 }
@@ -23,7 +23,7 @@ func newGroupFrame(contents ...shape.Bounded) Frame {
 	return groupFrame{Group: shape.NewGroup(contents...)}
 }
 
-func (f groupFrame) Translate(x, y float32) Frame {
+func (f groupFrame) Translate(x, y float64) Frame {
 	g := shape.NewGroup(f.Elements...).Translate(x, y)
 	return groupFrame{Group: g}
 }
